+++ src/client/components/ReactClient.tsx
@@ -1,6 +1,6 @@
 import { RelayOnlineClient_C, RelayOnlineClient_D, RelayOnlineClient_E, RelayClientWithBot_C, RelayClientWithBot_D, RelayClientWithBot_E } from "../../games/relay/main";
 import { OnlineClient_C as StrategyOnlineClient_C, OnlineClient_D as StrategyOnlineClient_D, OnlineClient_E as StrategyOnlineClient_E, ClientWithBot_C as StrategyClientWithBot_C, ClientWithBot_D as StrategyClientWithBot_D, ClientWithBot_E as StrategyClientWithBot_E,  } from "../../games/remove-from-circle/main";
-import { IS_OFFLINE_MODE } from "../utils/appMode";
+import { IS_OFFLINE_MODE } from "../utils/util";
 
 const DURER_XVI_CLIENT_C_RELAY = RelayOnlineClient_C;
 const DURER_XVI_CLIENT_D_RELAY = RelayOnlineClient_D;
