--- src/client/utils/appMode.ts
@@ -1,1 +1,0 @@
-export const IS_OFFLINE_MODE = process.env.REACT_APP_WHICH_VERSION === "b"
