+++ src/games/remove-from-circle/board.tsx
@@ -2,7 +2,7 @@ import { MyGameState } from './game';
 import { BoardProps } from 'boardgame.io/react';
 import { sendDataStrategyStep } from '../../common/sendData';
 import { range } from "lodash";
-import { IS_OFFLINE_MODE } from '../../client/utils/appMode';
+import { IS_OFFLINE_MODE } from '../../client/utils/util';
 
 
 interface MyGameProps extends BoardProps<MyGameState> { }
